Delete config_handler and replace with renamed config_handler_catalog to give user the choice of using a catalog to input parameter values. Fix seed values
